split out TargetFont class to its own file
and helper functions of utils.py (issue #1204)
